+++ platform/linuxbsd/x11/detect_prime_x11.cpp
