+++ platform/linuxbsd/x11/gl_manager_x11.cpp
